made the image drawing methods operate on graphics instead of painteventargs moved helper methods to extensions
